requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.14.0
fsspec==2024.3.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
requests==2.31.0
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
